refactor: Create `autocompleteLanguage()` fragment.
